% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_functions.R
\name{read.qPCR}
\alias{read.qPCR}
\title{Read a qPCR file}
\usage{
read.qPCR(
  file,
  sheet = "Results",
  results.range,
  columns = c("Sample Name", "Target Name", "CT", "Ct Mean"),
  undet.ct = "Undetermined"
)
}
\arguments{
\item{file}{pathway to the xls/xlsx file which the data are to be read from. Each row of the table appears as one line of the file. If it does not contain an absolute path, the file name is relative to the current working directory}

\item{sheet}{string name of excel's sheet where results are stored. If not specified, the default is "Results"}

\item{results.range}{An Excel cell range to read from}

\item{columns}{Character-based vector with the names of the columns to read from. By default column's named "Sample Name", "Target Name", "CT", "Ct Mean" are read}

\item{undet.ct}{string used by qPCR machine to name undetermined CT values. If not specified, the default value is "Undetermined", which ensures "CT" fields does not coerce as a character type.}
}
\value{
A dataframe with parsed data according to the variables wirtten in the 'columns' argument
}
\description{
Reads an Excel (.xls) file with the results from a qPCR with the data being parsed as asked in the 'columns' argument.
}
