% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qpcr_functions.R
\name{qPCR.results}
\alias{qPCR.results}
\title{Analyzes results from a qPCR}
\usage{
qPCR.results(
  qPCR_table,
  control_cond,
  hk_genes,
  rows = 1:nrow(qPCR_table),
  sample_name = "Sample Name",
  target_name = "Target Name",
  ct = "CT",
  ct_mean = "Ct Mean",
  include.ct_mean = TRUE,
  minus.ctrl = TRUE,
  minus.mean = TRUE
)
}
\arguments{
\item{qPCR_table}{An R data frame with the results of the qPCR, such as one generated by `read.qPCR()`.}

\item{control_cond}{A character indicating the condition to be considered as control or reference.}

\item{hk_genes}{A character vector indicating the name of the housekeeping genes.}

\item{rows}{A numeric vector indicating the index of the rows with samples belonging to the same experiment. By default, all rows are considered to belong to the same experiment (`1:nrow(qPCR_table)`).}

\item{sample_name}{Character indicating the name of the variable that stores sample names, `"Sample Name"` by default.}

\item{target_name}{Character indicating the name of the variable that stores target (gene) names, `"Target Name"` by default.}

\item{ct}{Character indicating the name of the variable that stores Ct values, `"CT"` by default.}

\item{include.ct_mean}{A logical value indicating whether Ct mean values (mean of Ct values of replicates) are included in the `qPCR_table` data frame, `TRUE` by default.}

\item{minus.ctrl}{A logical value indicating whether  ∆Ct,  ∆∆Ct, fold change and LFC are to be calculated with respect to a control or reference condition, `TRUE` by default.}

\item{minus.mean}{a logical value indicating whether  ∆Ct,  ∆∆Ct, fold change and LFC are to be calculated with respect to the mean value for same genes, `TRUE` by default.}

\item{ct_Mean}{Character indicating the name of the variable that stores Ct mean values (mean of Ct values of replicates), `"Ct Mean"` by default.}
}
\value{
A dataframe with ∆Ct and ∆∆Ct values according to the reference sample indicated and the average; and both fold change (2^-∆∆Ct) and LFC (log2 fold change)
}
\description{
Takes the results from a qPCR dataframe and returns their ∆Ct and ∆∆Ct values
according to a reference and the average; and both fold change and Log Fold Change
}
